Some camera collision + HUD changes
Started testing on mobile so I can make more additions to the level. Some of the HUD placements were difficult to play (Jump being right beside movement), so I reorganized some of it, changed some of the size since it was rather large.

Camera collision was fixed a bit, though I still need to test it some more